Cleaning the scripts
