+++ Assets/Resources/EnemyTrigger.prefab
@@ -14,6 +14,7 @@ GameObject:
   - component: {fileID: 5228476996973053668}
   - component: {fileID: 5228476996973053669}
   - component: {fileID: 5228476996973053670}
+  - component: {fileID: -724539745537424008}
   m_Layer: 0
   m_Name: EnemyTrigger
   m_TagString: Untagged
@@ -129,6 +130,21 @@ Rigidbody:
   m_Interpolate: 0
   m_Constraints: 80
   m_CollisionDetection: 0
+--- !u!114 &-724539745537424008
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 5228476996973053671}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: edad782c7e3e0cf47a408bcc08842739, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  colorChangeInterval: 0.5
+  glowColor: {r: 1, g: 0.92156863, b: 0.015686275, a: 1}
+  glowIntensity: 2.5
 --- !u!1 &6107301821116421247
 GameObject:
   m_ObjectHideFlags: 0
