+++ Assets/Resources/slapper.prefab
@@ -592,7 +592,6 @@ GameObject:
   - component: {fileID: 8060312172912194147}
   - component: {fileID: -4757617431728341033}
   - component: {fileID: -6504410244229579597}
-  - component: {fileID: -6333865540505204216}
   - component: {fileID: 1328788441910454822}
   m_Layer: 0
   m_Name: slapper
@@ -662,6 +661,9 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 845091760ea79ae47b19da03c2025e57, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  rotationSpeed: 180
+  spiralRadius: 25
+  spiralSpeed: 0
   pointA: {fileID: 3415095860265940418, guid: a4f22aa0dea4d9f4a9b4ca8674f4b72a, type: 3}
   pointB: {fileID: 6412056012148655471, guid: 8e839c091234ad346b4bcb7349f36d24, type: 3}
   pointC: {fileID: 8006011809170404262, guid: 39434ba7c6d15134cbc1136e80e8722d, type: 3}
@@ -687,22 +689,6 @@ MonoBehaviour:
   nextDronePrefab: {fileID: 3581606126056124790, guid: e4322f0e1412dae44b3de21aace6226d,
     type: 3}
   prefabsArray: []
---- !u!54 &-6333865540505204216
-Rigidbody:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 8060312172912194145}
-  serializedVersion: 2
-  m_Mass: 1
-  m_Drag: 0
-  m_AngularDrag: 0.05
-  m_UseGravity: 0
-  m_IsKinematic: 0
-  m_Interpolate: 0
-  m_Constraints: 80
-  m_CollisionDetection: 0
 --- !u!82 &1328788441910454822
 AudioSource:
   m_ObjectHideFlags: 0
