+++ Assets/Scripts/AI Enemies/Enemy Scripts/Floating enemy trigger.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: bfb040c6165125a4596b58457c117c78
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
