+++ Assets/Scripts/AI Enemies/Enemy Scripts/Floating enemy trigger/FloatingAITrigger.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8fbe4f707c907af4cb23427cbe80680e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
