+++ Assets/Scripts/AI Enemies/Enemy Scripts/Floating enemy trigger/ResizeObject.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6bd9a6c66f53552408c5f24db10824fa
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
