+++ Assets/Scripts/Weapons/LeftAndRightHand.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9a7c46266f4643a49a9336e295c78ad7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
